2023-03-mitmproxy-ga-checker/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.2.0
autopep8==2.0.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver-binary==111.0.5563.41.0
click==8.1.3
cryptography==38.0.4
Flask==2.2.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.2
mitmproxy==9.0.1
mitmproxy_wireguard==0.1.20
msgpack==1.0.4
outcome==1.2.0
passlib==1.7.4
protobuf==4.22.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
ruamel.yaml==0.17.21
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
urwid==2.1.2
Werkzeug==2.2.3
wsproto==1.2.0
zstandard==0.19.0
